+++ checksum-Elixir.SvgHush.Native.exs
@@ -1,0 +1,6 @@
+%{
+  "libsvg_hush_nif-v0.1.0-nif-2.17-aarch64-unknown-linux-gnu.so.tar.gz" =>
+    "sha256:4918c367f461d21c7edc928ae558effda1f6480eaa3fe3514cc804c071d6c301",
+  "libsvg_hush_nif-v0.1.0-nif-2.17-x86_64-unknown-linux-gnu.so.tar.gz" =>
+    "sha256:d7a61b3882120247e6429f392891c8c310d84a52fd346c2e49d90d8d1da33f9f"
+}